Fix example viewer reset discarding user CLI args
The example browser's reset path constructed a fresh parser and called
default_args() on it, so any flags the user passed on the command line
(e.g. --world-count 2) were thrown away and the example was rebuilt
with parser defaults.

Capture the original args namespace in _ExampleBrowser and reuse a
copy of it on reset, falling back to parser defaults only when no
args were supplied.

Also rename the manual GL smoke-test to smoketest_example_browser.py
so the unittest runner does not auto-discover it, and add real
regression tests under newton/tests/test_example_browser.py.

Fixes #2695.
